Allow pulling images from private Docker registry (#71)
* Allow pulling images from private Docker registry

As per the documentation for Fargate using private docker registries[1]
this commit introduces a new variable
`repository_credentials_secret_arn` used to pass in the AWS Secrets
Manager path to your credentials secret. Note that this also requires
the Lambda function execution task role to have access to this secret.

[1] https://aws.amazon.com/blogs/compute/introducing-private-registry-authentication-support-for-aws-fargate/

* Output the ECS task execution role name

We need the role name rather than the ARN of the role to attach policies
to it later. Yes we could infer the name from the ARN using split etc
but meh, just output it.

* Add `initProcessEnabled` to container def

Enable running container definitions with the —init flag to ensure a
proper supervisor runs our command.